pyproject.toml
[project]
name = "compliance-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "chromadb>=1.5.5",
    "datasets>=4.7.0",
    "diskcache>=5.6.3",
    "faiss-cpu>=1.13.2",
    "fastapi>=0.135.1",
    "fastembed>=0.7.4",
    "groq>=0.37.1",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-classic>=1.0.3",
    "langchain-cohere>=0.5.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-qdrant>=1.1.0",
    "openai>=2.28.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pymupdf>=1.27.2",
    "pymupdf4llm>=1.27.2.1",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "python-slugify>=8.0.4",
    "qdrant-client>=1.17.1",
    "ragas>=0.4.3",
    "sentence-transformers>=5.3.0",
    "supabase>=2.28.2",
    "tiktoken>=0.12.0",
    "uvicorn[standard]>=0.41.0",
]
